fix: extract svg handling functionality to separate hook
